Create folder for historical storm and seperate component for RenderSearchResult and renderRecentStorm
